% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nostep.R
\name{mh_mcmc_sir_nostep}
\alias{mh_mcmc_sir_nostep}
\title{Metropolis-Hastings MCMC for SIR (estimate I0 and β)}
\usage{
mh_mcmc_sir_nostep(
  cases,
  N,
  gamma,
  rho = 1,
  likelihood = c("poisson", "nbinom"),
  phi = NULL,
  n_iter = 20000,
  burn_in = 5000,
  thin = 10,
  init = list(I0 = 10, beta = 0.4),
  prop_sd = list(logI0 = 0.3, logbeta = 0.2),
  priors = list(meanlog_I0 = log(10), sdlog_I0 = 1, meanlog_beta = log(0.4), sdlog_beta =
    1),
  dt = 1,
  seed = NULL,
  verbose = TRUE
)
}
\description{
Metropolis-Hastings MCMC for SIR (estimate I0 and β)
}
