% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_nostep.R
\name{sir_nostep}
\alias{sir_nostep}
\title{Simulate SIR Model with Single Time-Varying Transmission Rate}
\usage{
sir_nostep(N, I0, gamma, beta, end_time)
}
\arguments{
\item{N}{Total population size}

\item{I0}{Initial number of infected individuals}

\item{gamma}{Recovery rate (1/infectious period)}

\item{beta}{Transmission rate before change point}

\item{end_time}{End time of simulation period}
}
\value{
A list containing:
\item{times}{Time points}
\item{S}{Susceptible counts over time}
\item{I}{Infected counts over time}
\item{R}{Recovered counts over time}
\item{beta}{Transmission rate over time}
\item{parameters}{List of parameters used}
}
\description{
Simulate SIR Model with Single Time-Varying Transmission Rate
}
