% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_nostep.R, R/sir_onestep.R
\name{sir_onestep}
\alias{sir_onestep}
\title{Simulate SIR Model with Single Time-Varying Transmission Rate}
\usage{
sir_onestep(N, I0, gamma, beta1, beta2, change_time, end_time)

sir_onestep(N, I0, gamma, beta1, beta2, change_time, end_time)
}
\arguments{
\item{N}{Total population size}

\item{I0}{Initial number of infected individuals}

\item{gamma}{Recovery rate (1/infectious period)}

\item{beta1}{Transmission rate before change point}

\item{beta2}{Transmission rate after change point}

\item{change_time}{Change point of transmission rate}

\item{beta}{Transmission rate before change point}

\item{times}{Vector of times at which to return output (default: 0 to 100 by 1)}
}
\value{
A list containing:
\item{times}{Time points}
\item{S}{Susceptible counts over time}
\item{I}{Infected counts over time}
\item{R}{Recovered counts over time}
\item{beta}{Transmission rate over time}
\item{parameters}{List of parameters used}

A list containing:
\item{times}{Time points}
\item{S}{Susceptible counts over time}
\item{I}{Infected counts over time}
\item{R}{Recovered counts over time}
\item{beta}{Transmission rate over time}
\item{parameters}{List of parameters used}
}
\description{
Simulate SIR Model with Single Time-Varying Transmission Rate

Simulate SIR Model with Single Time-Varying Transmission Rate
}
